Refactor: move sync widget tests
