+++ test/logic/persistence_logic_test.dart
@@ -21,9 +21,9 @@ import 'package:mocktail/mocktail.dart';
 
 import '../helpers/path_provider.dart';
 import '../mocks.dart';
-import '../sync/sync_config_test_data.dart';
-import '../sync/sync_config_test_mocks.dart';
+import '../mocks/sync_config_test_mocks.dart';
 import '../test_data.dart';
+import '../test_data/sync_config_test_data.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
