+++ test/widgets/sync/outbox_badge_test.dart
@@ -5,8 +5,8 @@ import 'package:lotti/get_it.dart';
 import 'package:lotti/pages/settings/outbox/outbox_badge.dart';
 import 'package:lotti/themes/themes_service.dart';
 
+import '../../mocks/sync_config_test_mocks.dart';
 import '../../widget_test_utils.dart';
-import '../sync_config_test_mocks.dart';
 
 void main() {
   var mock = MockSyncDatabase();
