+++ test/widgets/sync/qr_widget_test.dart
@@ -7,9 +7,9 @@ import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/widgets/sync/qr_widget.dart';
 import 'package:mocktail/mocktail.dart';
 
+import '../../mocks/sync_config_test_mocks.dart';
+import '../../test_data/sync_config_test_data.dart';
 import '../../widget_test_utils.dart';
-import '../sync_config_test_data.dart';
-import '../sync_config_test_mocks.dart';
 
 void main() {
   group('SyncConfig QR Widget Tests - ', () {
